Refactor Out Remote Proxy Exception
