+++ vcell-client/src/main/java/cbit/vcell/client/LocalDataSetControllerProvider.java
@@ -10,6 +10,7 @@
 
 package cbit.vcell.client;
 
+import cbit.vcell.message.server.bootstrap.client.RemoteProxyException;
 import org.vcell.solver.nfsim.NFSimMolecularConfigurations;
 import org.vcell.util.DataAccessException;
 import org.vcell.util.document.TimeSeriesJobResults;
@@ -25,7 +26,6 @@ import cbit.vcell.export.server.ExportServiceImpl;
 import cbit.vcell.export.server.ExportSpecs;
 import cbit.vcell.field.io.FieldDataFileOperationResults;
 import cbit.vcell.field.io.FieldDataFileOperationSpec;
-import cbit.vcell.message.server.bootstrap.client.RemoteProxyVCellConnectionFactory.RemoteProxyException;
 import cbit.vcell.server.DataSetController;
 import cbit.vcell.server.DataSetControllerProvider;
 import cbit.vcell.simdata.DataIdentifier;
