+++ vcell-client/src/main/java/cbit/vcell/client/VCellClientMain.java
@@ -15,7 +15,7 @@ import cbit.util.xml.XmlUtil;
 import cbit.vcell.client.desktop.NetworkProxyPreferences;
 import cbit.vcell.client.server.ClientServerInfo;
 import cbit.vcell.client.server.DynamicClientProperties;
-import cbit.vcell.message.server.bootstrap.client.RemoteProxyVCellConnectionFactory;
+import org.vcell.api.messaging.RemoteProxyVCellConnectionFactory;
 import cbit.vcell.mongodb.VCMongoMessage;
 import cbit.vcell.resource.*;
 import cbit.vcell.xml.XmlHelper;
@@ -23,7 +23,6 @@ import com.google.inject.Guice;
 import com.google.inject.Injector;
 import org.jdom.Document;
 import org.vcell.dependency.client.VCellClientModule;
-import org.vcell.util.document.UserLoginInfo;
 import org.vcell.util.document.VCDocument;
 import org.vcell.util.document.VCellSoftwareVersion;
 import org.vcell.util.logging.ConsoleCapture;
