+++ vcell-core/src/main/java/cbit/vcell/message/server/bootstrap/client/RemoteProxyException.java
@@ -1,0 +1,7 @@
+package cbit.vcell.message.server.bootstrap.client;
+
+public class RemoteProxyException extends Exception {
+    public RemoteProxyException(String message, Throwable cause) {
+        super(message, cause);
+    }
+}
