+++ vcell-core/src/main/java/cbit/vcell/server/SimulationController.java
@@ -9,10 +9,10 @@
  */
 
 package cbit.vcell.server;
+import cbit.vcell.message.server.bootstrap.client.RemoteProxyException;
 import org.vcell.util.DataAccessException;
 import org.vcell.util.document.KeyValue;
 
-import cbit.vcell.message.server.bootstrap.client.RemoteProxyVCellConnectionFactory.RemoteProxyException;
 import cbit.vcell.solver.VCSimulationIdentifier;
 /**
  * This interface was generated by a SmartGuide.
