+++ vcell-core/src/main/java/cbit/vcell/server/VCellBootstrap.java
@@ -10,13 +10,13 @@
 
 package cbit.vcell.server;
 
+import cbit.vcell.message.server.bootstrap.client.RemoteProxyException;
 import org.vcell.util.AuthenticationException;
 import org.vcell.util.DataAccessException;
 import org.vcell.util.UseridIDExistsException;
 import org.vcell.util.document.UserInfo;
 import org.vcell.util.document.UserLoginInfo;
 
-import cbit.vcell.message.server.bootstrap.client.RemoteProxyVCellConnectionFactory.RemoteProxyException;
 /**
  * This interface was generated by a SmartGuide.
  *
