+++ vcell-core/src/main/java/cbit/vcell/server/VCellConnection.java
@@ -10,15 +10,13 @@
 
 package cbit.vcell.server;
 
-import java.io.IOException;
-import java.io.Serializable;
-
+import cbit.rmi.event.MessageEvent;
+import cbit.vcell.message.server.bootstrap.client.RemoteProxyException;
 import org.vcell.util.DataAccessException;
 import org.vcell.util.document.UserLoginInfo;
 
-import cbit.rmi.event.MessageEvent;
-import cbit.rmi.event.PerformanceMonitorEvent;
-import cbit.vcell.message.server.bootstrap.client.RemoteProxyVCellConnectionFactory.RemoteProxyException;
+import java.io.IOException;
+import java.io.Serializable;
 /**
  * This interface was generated by a SmartGuide.
  *
