+++ vcell-core/src/main/java/org/vcell/service/registration/RegistrationService.java
@@ -1,11 +1,10 @@
 package org.vcell.service.registration;
 
+import cbit.vcell.message.server.bootstrap.client.RemoteProxyException;
 import org.vcell.util.DataAccessException;
 import org.vcell.util.UseridIDExistsException;
 import org.vcell.util.document.KeyValue;
 import org.vcell.util.document.UserInfo;
-
-import cbit.vcell.message.server.bootstrap.client.RemoteProxyVCellConnectionFactory.RemoteProxyException;
 
 public interface RegistrationService {
 	public UserInfo insertUserInfo(UserInfo newUserInfo,boolean bUpdate) throws RemoteProxyException,DataAccessException,UseridIDExistsException;
