+++ vcell-vmicro/src/main/java/org/vcell/vmicro/op/display/DisplayTimeSeriesOp.java
@@ -9,6 +9,7 @@ import java.lang.reflect.InvocationTargetException;
 import javax.swing.JFrame;
 import javax.swing.SwingUtilities;
 
+import cbit.vcell.message.server.bootstrap.client.RemoteProxyException;
 import org.vcell.client.logicalwindow.LWTopFrame;
 import org.vcell.solver.nfsim.NFSimMolecularConfigurations;
 import org.vcell.util.DataAccessException;
@@ -49,7 +50,6 @@ import cbit.vcell.field.io.FieldDataFileOperationSpec;
 import cbit.vcell.geometry.RegionImage;
 import cbit.vcell.math.Variable.Domain;
 import cbit.vcell.math.VariableType;
-import cbit.vcell.message.server.bootstrap.client.RemoteProxyVCellConnectionFactory.RemoteProxyException;
 import cbit.vcell.server.DataSetController;
 import cbit.vcell.server.DataSetControllerProvider;
 import cbit.vcell.simdata.ClientPDEDataContext;
